can delete elections with data

+++ js/election_list/election_list.js
@@ -30,6 +30,8 @@ angular.module('ElectionListModule', [
                 election = response.data.updated;
                 election.label = '{0}-{1}'.format(
                   election.start_time.split('-')[0], election.name);
+                election.candidates = [];
+                election.voted = 0;
                 scope.elections.push(election);
                 scope.currentElection = election;
               } else {
@@ -75,10 +77,16 @@ angular.module('ElectionListModule', [
         };
   
         scope.deleteElection = (election) => {
+          if (!confirm('请确认您要删除"{0}"的全部数据'
+              .format(election.label))) {
+            return;
+          }
           rpc.delete_election(election.id).then((response) => {
             if (response.data.deleted) {
               var index = scope.elections.indexOf(election);
               scope.elections.splice(index, 1);
+              election.deleted = 1;
+              election.candidates.length = 0;
               if (election == scope.currentElection) {
                 scope.currentElection = null;
               }
